+++ lib/plugins/aws/package/compile/events/cloudFront.js
@@ -3,7 +3,7 @@
 const _ = require('lodash');
 const url = require('url');
 const chalk = require('chalk');
-const { ServerlessError } = require('../../../../../../classes/Error');
+const { ServerlessError } = require('../../../../../classes/Error');
 
 const originLimits = { maxTimeout: 30, maxMemorySize: 10240 };
 const viewerLimits = { maxTimeout: 5, maxMemorySize: 128 };
