+++ lib/plugins/index.js
@@ -50,7 +50,7 @@ module.exports = [
   require('./aws/package/compile/events/cognitoUserPool/index.js'),
   require('./aws/package/compile/events/eventBridge/index.js'),
   require('./aws/package/compile/events/sqs/index.js'),
-  require('./aws/package/compile/events/cloudFront/index.js'),
+  require('./aws/package/compile/events/cloudFront.js'),
   require('./aws/package/compile/events/httpApi'),
   require('./aws/deployFunction.js'),
   require('./aws/deployList.js'),
