+++ test/unit/lib/plugins/aws/package/compile/events/cloudFront.test.js
@@ -2,10 +2,10 @@
 
 const chai = require('chai');
 const sandbox = require('sinon');
-const AwsProvider = require('../../../../../../../../../lib/plugins/aws/provider/awsProvider');
-const AwsCompileCloudFrontEvents = require('../../../../../../../../../lib/plugins/aws/package/compile/events/cloudFront/index');
-const Serverless = require('../../../../../../../../../lib/Serverless');
-const runServerless = require('../../../../../../../../utils/run-serverless');
+const AwsProvider = require('../../../../../../../../lib/plugins/aws/provider/awsProvider');
+const AwsCompileCloudFrontEvents = require('../../../../../../../../lib/plugins/aws/package/compile/events/cloudFront');
+const Serverless = require('../../../../../../../../lib/Serverless');
+const runServerless = require('../../../../../../../utils/run-serverless');
 
 chai.use(require('sinon-chai'));
 chai.use(require('chai-as-promised'));
